note ignores that lifx_legacy was removed 12 months ago
I'm a new (windows/vm) HA user fumbling my way around and wasted a lot of time trying to find this lifx_legacy thing, before discovering it was removed 12 months ago. I'm not yet aware if this means windows support is non-existent or (hopefully) it DOES now work with the standard LIFX integration, but I think this note about using the legacy component needs to be amended or removed.